fix: refactoring solver as a submodule
