+++ src/solver/__init__.py
@@ -1,0 +1,1 @@
+from .algorithms.ucs import UCSSolver
