+++ src/solver/base.py
